Further fix issues with movement.
For future reference, the issue was that I was getting 'lucky' with the
cube being symmetrical/having a column/row being the same color. So if
the row/column was reversed/flipped in any manner, I didn't notice.

I am going to have write unit tests to thoroughly debug this model
before moving on to the solution.